VIDSOL-867 add feedback buttons
