+++ VERA/VERACommonUI/VERACommonUI/Resources/Images.xcassets/feedback-line.imageset/Contents.json
@@ -8,5 +8,8 @@
   "info" : {
     "author" : "xcode",
     "version" : 1
+  },
+  "properties" : {
+    "template-rendering-intent" : "template"
   }
 }
