--- VERA/VERAFeedback/VERAFeedback/UI/View models/FeedbackViewModel.swift
@@ -1,15 +1,0 @@
-//
-//  Created by Vonage on 3/6/26.
-//
-
-import Foundation
-
-/// View model for FeedbackView.
-@MainActor
-public final class FeedbackViewModel: ObservableObject {
-    // TODO: Add published properties and dependencies
-
-    public init() {
-        // TODO: Add initialisation, inject usecase
-    }
-}
