+++ VERA/VERAMeetingRoomSDK/VERAMeetingRoomSDK/Resources/Localizable.xcstrings
@@ -37,6 +37,10 @@
         }
       }
     },
+    "Feedback" : {
+      "comment" : "Label for the feedback button.",
+      "isCommentAutoGenerated" : true
+    },
     "Noise Suppression" : {
       "comment" : "Label for the audio effects button.",
       "isCommentAutoGenerated" : true,
